+++ src/test/java/com/teragrep/akv_01/event/JSONEventTest.java
@@ -73,7 +73,7 @@ public final class JSONEventTest {
         Assertions.assertEquals(json.toString(), event.asString());
         Assertions.assertEquals("123", event.resourceId());
         Assertions.assertEquals(json, event.asJsonStructure());
-        Assertions.assertEquals(ZonedDateTime.parse("2010-01-01T00:00Z"), event.enqueuedTime().parsed());
+        Assertions.assertEquals(ZonedDateTime.parse("2010-01-01T00:00Z"), event.enqueuedTime().zonedDateTime());
         Assertions.assertTrue(event.isJsonStructure());
     }
